requirements.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
aiofiles==24.1.0
psycopg2-binary==2.9.10
openai==1.14.3
python-docx==1.1.0
PyPDF2==3.0.1
python-magic==0.4.27

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
authlib==1.3.0
httpx==0.25.2

# Validation
pydantic==2.5.2
pydantic-settings==2.1.0
email-validator==2.1.0

# CORS
fastapi-cors==0.0.6

# Testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1

# Development
black==23.12.0
isort==5.13.2

# Report Generation
reportlab==4.0.7
pandas==2.1.4
jinja2==3.1.3